+++ api/v1/views/__init__.py
@@ -19,6 +19,6 @@ from .users import *
 from .places import *
 from .states import *
 from .cities import *
-from .reviews import *
 from .amenities import *
+from .places_reviews import *
 from .places_amenities import *
